[Maintenance] Rename builders to factories
